requirements.txt
pymongo~=4.5.0
paho-mqtt~=1.6.1
openpyxl~=3.1.2
certifi~=2023.7.22
pytz~=2023.3.post1
chardet~=5.2.0